@gustavogalli [Task: inserindo métodos de logar e cadastrar em UsuarioController]

+++ src/main/java/com/projetointegradorg3/redeSocial/controller/UsuarioController.java
@@ -1,6 +1,7 @@
 package com.projetointegradorg3.redeSocial.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -15,11 +16,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
 import com.projetointegradorg3.redeSocial.model.Usuario;
+import com.projetointegradorg3.redeSocial.model.UsuarioLogin;
 import com.projetointegradorg3.redeSocial.repository.UsuarioRepository;
+import com.projetointegradorg3.redeSocial.service.UsuarioService;
 
 @RestController
 @RequestMapping("/usuario")
@@ -28,32 +32,50 @@ import com.projetointegradorg3.redeSocial.repository.UsuarioRepository;
 public class UsuarioController {
 
 	@Autowired
+	private UsuarioService usuarioService;
+
+	@Autowired
 	private UsuarioRepository repository;
 
-	@GetMapping
-	public ResponseEntity<List<Usuario>> GetAll() {
+	@GetMapping("/all")
+	public ResponseEntity <List<Usuario>> getAll() {
 		return ResponseEntity.ok(repository.findAll());
 	}
 
 	@GetMapping("/{id}")
-	public ResponseEntity<Usuario> GetById(@PathVariable Long id) {
-		return repository.findById(id).map(resp -> ResponseEntity.status(200).body(resp))
-				.orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "Id não existente"));
+	public ResponseEntity<Usuario> getById(@PathVariable long id) {
+		return repository.findById(id)
+			.map(resp -> ResponseEntity.ok(resp))
+			.orElse(ResponseEntity.notFound().build());
 	}
 
-	@PostMapping("/save")
+	@PostMapping("/logar")
+	public ResponseEntity<UsuarioLogin> autentication(@RequestBody Optional<UsuarioLogin> usuario) {
+		return usuarioService.logarUsuario(usuario)
+			.map(resp -> ResponseEntity.ok(resp))
+			.orElse(ResponseEntity.status(HttpStatus.UNAUTHORIZED).build());
+	}
+
+	@PostMapping("/cadastrar")
 	public ResponseEntity<Usuario> post(@Valid @RequestBody Usuario usuario) {
-		return ResponseEntity.status(200).body(repository.save(usuario));
+		return usuarioService.cadastrarUsuario(usuario)
+			.map(resp -> ResponseEntity.status(HttpStatus.CREATED).body(resp))
+			.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
 
-	@PutMapping("/update")
-	public ResponseEntity<Usuario> put(@Valid @RequestBody Usuario usuario) {
-		return ResponseEntity.status(200).body(repository.save(usuario));
+	@PutMapping("/atualizar")
+	public ResponseEntity<Usuario> put(@Valid @RequestBody Usuario usuario){
+		return usuarioService.atualizarUsuario(usuario);
 	}
+	
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	@DeleteMapping("/{id}")
+	public void delete(@PathVariable long id) {
+		Optional<Usuario> usuario = repository.findById(id);
 
-	@DeleteMapping(path = "/{id}")
-	public void delete(@PathVariable Long id) {
+		if(usuario.isEmpty())
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
 		repository.deleteById(id);
 	}
-
 }
